requirements.txt
tensorflow>=2.15.0
numpy>=1.24.0
python-Levenshtein>=0.23.0
jupyter>=1.0.0
notebook>=7.0.0
ipykernel>=6.0.0 